Move test to failing directory
